requirements.txt
# Core dependencies
numpy
tqdm
tenacity
openai
tiktoken

# PDF processing libraries (at least one is required)
PyMuPDF
pdfplumber
PyPDF2

# Optional: OCR support (for scanned PDFs)
pillow
pytesseract

#Utilities
python-dotenv
cdifflib
statistics